requirements.txt
seaborn==0.12.2
networkx==3.1
wordcloud==1.9.2
pandas==1.5.3
scipy==1.10.1
matplotlib-venn==0.11.9
selenium==3.141.0
beautifulsoup4==4.12.2
google-api-python-client==2.108.0
